requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# PDF Processing
PyMuPDF==1.23.26
pdfplumber==0.11.0
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows

# Text Processing & Embeddings
sentence-transformers==2.3.1
langchain-text-splitters==0.0.1
tiktoken==0.5.2
nltk==3.8.1

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3
joblib==1.3.2

# LLM APIs
# openai>=1.30.0  # Commented out - using Google Gemini instead
google-generativeai>=0.3.0
tenacity==8.2.3

# Voice Processing
pydub==0.25.1
SpeechRecognition==3.10.1

# Utilities
requests==2.31.0
python-multipart==0.0.6
validators==0.22.0
python-dateutil==2.8.2
Pillow==10.2.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0
ipython==8.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
